requirements.txt
pymongo>=4.6.0
click>=8.1.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

numpy>=1.24.0
faker>=20.0.0
colorama>=0.4.6

